app/requirements.txt
matplotlib==3.6.2
plotly==5.11.0
scipy==1.9.3
scikit-learn==1.2.0
bs4==0.0.1
openai==0.25.0
tiktoken==0.1.1
python-dotenv==0.21.0
Flask==2.2.2